requirements.txt
weasyprint>=61.0
genanki>=0.13.0
pyyaml>=6.0
wyoming>=1.5.4
requests>=2.31.0
pillow>=10.0.0
jinja2>=3.1.0
pydub>=0.25.1
cairosvg>=2.7.0
